requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
python-dotenv
